pyproject.toml
[tool.poetry]
name = "pluginforge"
version = "0.1.0"
description = "Application-agnostic plugin framework built on pluggy"
authors = ["Asterios Raptis"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/astrapi69/pluginforge"
keywords = ["plugin", "framework", "pluggy", "hooks", "yaml"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
packages = [{ include = "pluginforge" }]

[tool.poetry.dependencies]
python = "^3.11"
pluggy = "^1.4.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
ruff = "^0.4"

[tool.poetry.extras]
fastapi = ["fastapi"]
alembic = ["alembic", "sqlalchemy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=pluginforge --cov-report=term-missing"

[tool.ruff]
target-version = "py311"
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
